Make solutions be part of doctests
This means that the documentation will contain the solution code inline
with the solution explanation, and we can still test it using doctests.

However, this also means that `cargo run` will no longer work. This is
acceptable, since if you really want to run a specific solution locally
you can do `cargo test --doc problem_2` or whatever the problem number
is. I may eventually use `cargo make` for this.

This change also measures runtime of each solution and adds it to the
docs.