Improve the performance of solution 9
Previously, we were doing a couple things wrong. First, we were
iterating through all possible values of `c` even though once we have an
`a` and a `b` we can just subtract them to get `c`. Second, we didn't
need to iterate through values of `b` up to 1000 since `a + b` can't be
more than 1000.